pyproject.toml
[project]
name = "mcp-starter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "fastmcp>=2.11.2",
    "markdownify>=1.1.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.10",
    "langchain-openai>=0.3.0",
    "langgraph>=0.1.17",
    "langgraph-supervisor>=0.0.5",
    "langchain-google-genai>=0.1.0",
    "langchain-google-vertexai = ^0.1.0",
    "langchain-mcp-adapters>=0.1.2"
]
